Refactor implementation of to/fromJson methods on fields
The ConvertHelper class exposed in 1.3.0 is effectively worthless
outside of json_serializable because it relies on internal state that
assumes the in-box generator.

This commit exposes the hooks necessary for another generator to
use ConvertHelper by providing callbacks for providing ConvertData
for a given context.

Eliminates the internal JsonKeyWithConversion class
Exposes a new ConvertData class

Refactors the to/fromJson logic out of `json_key_with_conversion.dart`
into its own file - which should make maintenance easier