Upgrade to analyzer 0.39.x
Towards #552

Behavior related to type checks have been moved to `TypeSystem`. Keep an
instance of the type system available on `HelperCore` so that we can
access it synchronously where it is needed.

Remove an `assert` that can no longer be supported due to not having a
`TypeSystem` argument available. It would be breaking to change the
signature to pass one in.

Remove one conditional on equivalent types that can no longer be
performed without a breaking change. This likely also breaks behavior
and we will need to find a solution.