Add JsonSerializable.genericArgumentFactories
When `true` on classes with type parameters (generic types), extra
"helper" parameters will be generated for `fromJson` and/or `toJson` to
support serializing values of those types.

For example, the generated code for

```dart
@JsonSerializable(genericArgumentFactories: true)
class Response<T> {
  int status;
  T value;
}

```

Looks like

```dart
Response<T> _$ResponseFromJson<T>(
  Map<String, dynamic> json,
  T Function(Object json) helperForT,
) {
  return Response<T>()
    ..status = json['status'] as int
    ..value = helperForT(json['value']);
}
Map<String, dynamic> _$ResponseToJson<T>(
  Response<T> instance,
  Object Function(T value) helperForT,
) =>
    <String, dynamic>{
      'status': instance.status,
      'value': helperForT(instance.value),
    };
```

Note: this option has no effect on classes class without type parameters.
A warning is printed in such cases.

Note: if this option is set for all classes in a package via `build.yaml`
it is only applied to classes with type parameters – so no warning is
printed.

Related to

https://github.com/google/json_serializable.dart/issues/593
https://github.com/google/json_serializable.dart/issues/599
https://github.com/google/json_serializable.dart/issues/646
https://github.com/google/json_serializable.dart/issues/671